clear signature after changing account
