+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "src",
-  "version": "0.1.0+11.3",
+  "version": "0.1.0+11.4",
   "private": true,
   "scripts": {
     "dev": "next dev",
